+++ src/gui/editors/AutomationEditor.cpp
@@ -1074,8 +1074,7 @@ void AutomationEditor::paintEvent(QPaintEvent * pe )
 		else
 		{
 			int level = (int) m_bottomLevel;
-			int printable = qMax( 1, 5 * DEFAULT_Y_DELTA
-								/ m_y_delta );
+			int printable = qMax(1.0f, 5 * DEFAULT_Y_DELTA / m_y_delta);
 			int module = level % printable;
 			if( module )
 			{
@@ -1884,7 +1883,7 @@ void AutomationEditor::updateYDelta()
 	else
 	{
 		int grid_bottom = height() - SCROLLBAR_SIZE - 1;
-		m_y_delta = (grid_bottom - TOP_MARGIN) / (m_maxLevel - m_minLevel);
+		m_y_delta = static_cast<float>(grid_bottom - TOP_MARGIN) / (m_maxLevel - m_minLevel);
 	}
 }
 
@@ -1919,22 +1918,22 @@ void AutomationEditor::updateTopBottomLevels()
 		int centralLevel = (int)( m_minLevel + m_maxLevel - m_scrollLevel );
 
 		m_bottomLevel = centralLevel - ( half_grid
-							/ (float)m_y_delta );
+							/ m_y_delta );
 		if( m_bottomLevel < m_minLevel )
 		{
 			m_bottomLevel = m_minLevel;
 			m_topLevel = m_minLevel + (int)floorf( grid_height
-							/ (float)m_y_delta );
+							/ m_y_delta );
 		}
 		else
 		{
 			m_topLevel = m_bottomLevel + (int)floorf( grid_height
-							/ (float)m_y_delta );
+							/ m_y_delta );
 			if( m_topLevel > m_maxLevel )
 			{
 				m_topLevel = m_maxLevel;
 				m_bottomLevel = m_maxLevel - (int)floorf(
-					grid_height / (float)m_y_delta );
+					grid_height / m_y_delta );
 			}
 		}
 	}
